Started BottomPanel, finished Menu, added Images
Started writting BottomPanel class and its to-be children (DicePanel and
CardPanel). The Menu class is pretty much finished, I added one tool tip
to see how it works, but I will add more later.
I added the images required for the GUI, that is the 6 dice faces, and
the weapon, room, and character cards.