GUI almost complete, I think.
"BottomPanel" was renamed to "InfoPanel",  "DicePanel" was renamed to
"DiceDisplay" and now extends Component instead of JPanel, furthermore
it has been placed within another JPanel(LeftPanel), which will also
contain the "MessagePanel" and take up  1/3 of the InfoPanel, the
remainder of the InfoPanel will be taken up by the CardPanel. Almost all
the classes within the GUI package are mostly complete with the
exception of "MessagePanel" and the "BoardCanvas" which still needs an
action listener.
The GUI now has a method to display a hand of cards, and a method to
display two die with values between 1 and 6.