Implemented all the gameplay using GUI.
The program has now been adapted to use the GUI when a player takes
their turn.  To achieve this, a search algorithm had to be implemented
in board to find the best routes between tiles. There now a break
between players turns, where the next player must confirm that they wish
to begin their turn.
The game is now somewhat playable.